refactor: 패키지 위치 변경 (#253)
- 인터페이스는 domain, 구현체는 infrastructure에 위치
